refactor(app): review: implement-authentication
