--- apps/web/src/assets/images/index.ts
@@ -1,3 +1,0 @@
-import IbiriLogo from "./ibiri-logo.webp";
-
-export { IbiriLogo };
