+++ apps/web/src/components/RegisterForm.tsx
@@ -1,49 +1,29 @@
 'use client';
 
-import { useState } from 'react';
-import { useRouter } from 'next/navigation';
 import Image from 'next/image';
-import { IbiriLogo } from '../assets/images';
 import Link from 'next/link';
-import { useAuth } from '../context/AuthContext';
+import useRegister from '../hooks/use-register';
 
 export default function RegisterForm() {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [firstName, setFirstName] = useState('');
-  const [lastName, setLastName] = useState('');
-  const [error, setError] = useState('');
-  const router = useRouter();
-  const { login } = useAuth();
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setError('');
-
-    const response = await fetch('/api/register', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ email, password, firstName, lastName }),
-    });
-
-    const data = await response.json();
-
-    if (!response.ok) {
-      setError(data.error || 'Something went wrong');
-      return;
-    }
-
-    // Registration Success
-    login(data.user);
-    router.push('/');
-  };
+  const {
+    handleSubmit,
+    error,
+    firstName,
+    lastName,
+    email,
+    password,
+    setFirstName,
+    setLastName,
+    setEmail,
+    setPassword,
+  } = useRegister();
 
   return (
     <div className="flex min-h-screen flex-col items-center justify-center">
       <div className="w-full max-w-md space-y-8">
         <Image
           className="mix-blend-multiply"
-          src={IbiriLogo}
+          src="/images/ibiri-logo.webp"
           alt="Profile picture"
           width={500}
           height={500}
