+++ apps/web/src/components/withAuth.tsx
@@ -30,7 +30,7 @@ export function withAuth<P extends ComponentProps>(
     WrappedComponent.displayName || WrappedComponent.name || 'Component'
   })`;
 
-  return AuthComponent;
+  return AuthComponent as any;
 }
 
 export function withoutAuth<P extends ComponentProps>(
@@ -57,5 +57,5 @@ export function withoutAuth<P extends ComponentProps>(
     WrappedComponent.displayName || WrappedComponent.name || 'Component'
   })`;
 
-  return NoAuthComponent;
+  return NoAuthComponent as any;
 }
