+++ libs/db/src/lib/mongodb.ts
@@ -1,21 +1,20 @@
-import { MongoClient, ServerApiVersion } from 'mongodb';
+import { MongoClient, MongoClientOptions, ServerApiVersion } from 'mongodb';
 
-if (!process.env['NEXT_PUBLIC_MONGODB_URI']) {
-  throw new Error(
-    'Invalid/Missing environment variable: "NEXT_PUBLIC_MONGODB_URI"'
-  );
+if (!process.env['MONGODB_URI']) {
+  throw new Error('Invalid/Missing environment variable: "MONGODB_URI"');
 }
 
-const uri = process.env['NEXT_PUBLIC_MONGODB_URI'];
-const options = {
+const uri = process.env['MONGODB_URI'];
+const options: MongoClientOptions = {
   serverApi: {
     version: ServerApiVersion.v1,
     strict: true,
     deprecationErrors: true,
   },
+  appName: 'ibiri',
 };
 
-let client;
+let client: MongoClient;
 let clientPromise: Promise<MongoClient>;
 
 if (process.env['NODE_ENV'] === 'development') {
