+++ package.json
@@ -5,6 +5,7 @@
   "scripts": {},
   "private": true,
   "dependencies": {
+    "bcryptjs": "^2.4.3",
     "mongodb": "^6.9.0",
     "next": "14.2.3",
     "react": "18.3.1",
@@ -21,6 +22,7 @@
     "@swc-node/register": "~1.9.1",
     "@swc/core": "~1.5.7",
     "@swc/helpers": "~0.5.11",
+    "@types/bcryptjs": "^2.4.6",
     "@types/jest": "^29.5.12",
     "@types/node": "18.16.9",
     "@types/react": "18.3.1",
